+++ app/Service/Utilities/Calculator.php
@@ -4,7 +4,7 @@ namespace App\Service\Utilities;
 
 class Calculator
 {
-    public function sumBaseWithPercent(int $base, int $percentage): int
+    public function sumBaseWithPercent(int $base, float $percentage): int
     {
         return $base + floor(($percentage / 100) * $base);
     }
